Bump github.com/jackc/pgx/v5 from 5.8.0 to 5.9.2 (#56)
* Bump github.com/jackc/pgx/v5 from 5.8.0 to 5.9.2

Bumps [github.com/jackc/pgx/v5](https://github.com/jackc/pgx) from 5.8.0 to 5.9.2.
- [Changelog](https://github.com/jackc/pgx/blob/master/CHANGELOG.md)
- [Commits](https://github.com/jackc/pgx/compare/v5.8.0...v5.9.2)

---
updated-dependencies:
- dependency-name: github.com/jackc/pgx/v5
  dependency-version: 5.9.2
  dependency-type: direct:production
...


* fix: use pgx byte-based backend secret key in handshake

Agent-Logs-Url: https://github.com/TelenorNorway/azure-postgres-auth-proxy/sessions/5c2df745-b03c-43a1-a8f0-71138eb7324c


---------

+++ main.go
@@ -289,7 +289,7 @@ func simulateClientHandshake(clientConn net.Conn, hijackedBackend *pgconn.Hijack
 	}
 	handshakeMessages = append(handshakeMessages, &pgproto3.BackendKeyData{
 		ProcessID: uint32(hijackedBackend.PID),
-		SecretKey: uint32(hijackedBackend.SecretKey),
+		SecretKey: hijackedBackend.SecretKey,
 	})
 	handshakeMessages = append(handshakeMessages, &pgproto3.ReadyForQuery{
 		TxStatus: 'I',
